+++ packages/react-native-editor/CHANGELOG.md
@@ -11,6 +11,7 @@ For each user feature we should also add a importance categorization label  to i
 
 ## Unreleased
 -   [*] [Embed block] Fix inline preview cut-off when editing URL [#35321]
+-   [*] [Unsupported Block Editor] Fix text selection bug for Android [#34668]
 
 ## 1.63.0
 -   [**] [Embed block] Add the top 5 specific embed blocks to the Block inserter list [#34967]
@@ -22,7 +23,7 @@ For each user feature we should also add a importance categorization label  to i
 
 ## 1.62.1
 -   [**] Image block: fix height and border regression. [https://github.com/WordPress/gutenberg/pull/34957]
--   [**] Column block: fix width attribute flout cutoff. [https://github.com/wordpress-mobile/gutenberg-mobile/pull/3921]
+-   [**] Column block: fix width float attribute cut off. [#35061]
 
 ## 1.62.0
 -   [**] [Embed block] Implement WP embed preview component [#34004]